Create a dedicated cloud-init profile without install
Create a new Publiccloud cloud-init profile that does not contain
the `zypper in ed` directive.
Allow selecting which profile the test use via variable/setting
PUBLIC_CLOUD_CLOUD_INIT.
- Not defined --> no cloud-init deployment
- 'install' --> data/publiccloud/cloud-init-install.yaml
- any value --> data/publiccloud/cloud-init.yaml

+++ lib/publiccloud/instance.pm
@@ -657,7 +657,7 @@ sub cleanup_cloudinit() {
     $self->ssh_assert_script_run('sudo cloud-init clean --logs');
     if (get_var('PUBLIC_CLOUD_CLOUD_INIT')) {
         $self->ssh_assert_script_run('sudo rm /root/test_cloud-init.txt');
-        $self->ssh_assert_script_run('sudo zypper -n rm ed');
+        $self->ssh_assert_script_run('sudo zypper -n rm ed') if check_var('PUBLIC_CLOUD_CLOUD_INIT', 'install');
     }
 }
